requirements.txt
streamlit
tensorflow
numpy
pillow
streamlit-drawable-canvas
scipy
